Move non-location RHs to misc
